MAPSJS-2660: Simplify powerOfTwo generation.
Also, move old test to the right place.
