+++ @here/harp-geoutils/test/TileKeyUtilsTest.ts
@@ -4,10 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { expect } from "chai";
+import { assert, expect } from "chai";
 
+import { GeoBox } from "../lib/coordinates/GeoBox";
+import { GeoCoordinates } from "../lib/coordinates/GeoCoordinates";
 import { TileKey } from "../lib/tiling/TileKey";
 import { TileKeyUtils } from "../lib/tiling/TileKeyUtils";
+import { webMercatorTilingScheme } from "../lib/tiling/WebMercatorTilingScheme";
 
 describe("TileKeyUtils", function () {
     it("test getKeyForTileKeyAndOffset and extractOffsetAndMortonKeyFromKey", async function () {
@@ -47,5 +50,19 @@ describe("TileKeyUtils", function () {
             expect(offset).to.be.equal(offsetResults[i]);
             expect(mortonCode).to.be.equal(tileKey.mortonCode());
         }
+    });
+
+    it("geoRectangleToTileKeys", function () {
+        const geoBox = new GeoBox(
+            new GeoCoordinates(52.5163, 13.3777), // Brandenburg gate
+            new GeoCoordinates(52.5309, 13.385) // HERE office
+        );
+        const expectedResult = [371506848, 371506849, 371506850, 371506851];
+
+        const result = TileKeyUtils.geoRectangleToTileKeys(webMercatorTilingScheme, geoBox, 14);
+        assert.sameMembers(
+            result.map(tk => tk.mortonCode()),
+            expectedResult
+        );
     });
 });
